Moved modified dependencies to lib/
