+++ controllers/upload.js
@@ -2,7 +2,7 @@ const fs = require('fs')
 const path = require('path')
 const GIFEncoder = require('gif-stream/encoder')
 const GIFDecoder = require('gif-stream/decoder')
-const gifParser = require('../image_parsers/gif/gify.js')
+const gifParser = require('../lib/gify.js')
 const PNG = require('pngjs').PNG
 const settings = require('../settings.json')
 const stream = require('stream') // for turning Buffer to stream for the pipe API.
