+++ lib/cas-authentication.js
@@ -1,5 +1,5 @@
 
-const User        = require('./models/user.js')
+const User        = require('../models/user.js')
 var url           = require('url'),
     http          = require('http'),
     https         = require('https'),
