+++ lib/gify.js
